Register And Login Api Created

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('email')->unique();
-            $table->integer('phone');
-            $table->integer('age');
+            $table->bigInteger('phone');
+            $table->bigInteger('age');
             $table->longText('address');
             $table->string('password');
             $table->rememberToken();
